\name{get_IQR_condition_exp}
\alias{get_IQR_condition_exp}
\title{Calculate the IQR for each bin of the correlation matrix}
\usage{get_IQR_condition_exp(cor_mat, ave_logrpkm)}
\arguments{
\item{cor_mat}{Matrix, correlation matrix, generated by gene expression matrix, with genes sorted by average expression levels}
\item{ave_logrpkm}{Vector, average expression level of each gene}
}
\value{A matrix containing the IQR of each bin}
\description{This function allows users to compare the IQR of bins at different location of the correlation matrix.
}
\examples{
a=rnorm(10000)
b=a/max(abs(a))
cor_mat=array(b,dim=c(100,100))
ave_logrpkm=c(1:100)
get_IQR_condition_exp(cor_mat,ave_logrpkm)
}
