% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{normalize_correlation}
\alias{normalize_correlation}
\title{Spatial Quantile Normalization}
\usage{
normalize_correlation(cor_mat, ngrp, size_grp, ref_grp)
}
\arguments{
\item{cor_mat}{Matrix, correlation matrix, generated by sorted expression matrix}

\item{ngrp}{Integer, number of bins to use in the normalization}

\item{size_grp}{Integer, size of large bin}

\item{ref_grp}{Integer, the location of reference bin that represents target distribution}
}
\value{
The normalized correlation matrix
}
\description{
Spatial Quantile Normalization
}
\examples{
a=rnorm(10000)
b=a/max(abs(a))
cor_ori=array(b,dim=c(100,100))
normalize_correlation(cor_ori,ngrp=10,size_grp=15,ref_grp=9)

}
