\name{qqplot_condition_exp}
\alias{qqplot_condition_exp}
\title{Q-qplot for comparing the distribution of two bins in the correlation matrix}
\usage{qqplot_condition_exp(cor_mat, i,j)}
\arguments{
\item{cor_mat}{Matrix, correlation matrix, generated by gene expression matrix, with genes sorted by average expression levels}
\item{i}{Integer, number of the tested bin}
\item{j}{Integer, number of the reference bin}
}
\value{}
\description{This function allows users to compare the distribution of two bins using qqplot.
}
\examples{
a=rnorm(10000)
b=a/max(abs(a))
cor_mat=array(b,dim=c(100,100))
qqplot_condition_exp(cor_mat,1,9)
}
