\name{spqn-package}
\alias{spqn-package}
\alias{spqn}
\docType{package}
\title{
\packageTitle{spqn}
}
\description{
\packageDescription{spqn}
}
\details{
  See references for details on spatial quantile normalization.

  The main function is \code{normalize_correlation}. We include a number
  of plotting functions for examining the mean-correlation relationship,
  see the vignette for examples.
  
}
\references{
  Y Wang, SC Hicks, KD Hansen (2020).
  \emph{Co-expression analysis is biased by a mean-correlation relationship.}
  bioRxiv 2020.02.13.944777. \doi{10.1101/2020.02.13.944777}
}
\keyword{ package }
