\name{get_IQR_condition_exp}
\alias{get_IQR_condition_exp}
\title{Calculate the IQR for each bin of the correlation matrix}
\usage{get_IQR_condition_exp(cor_mat, ave_logrpkm)}
\arguments{
\item{cor_mat}{Matrix, correlation matrix, generated by gene expression matrix, with genes sorted by average expression levels}
\item{ave_logrpkm}{Vector, average expression level of each gene}
}
\value{A matrix containing the IQR of each bin}
\description{This function allows users to compare the IQR of bins at different location of the correlation matrix.
}
\examples{
  if(require(spqnData)) {
    data(gtex.4k)
    cor_mat <- cor(t(assay(gtex.4k)))
    ave_logrpkm <- rowData(gtex.4k)$ave_logrpkm
    get_IQR_condition_exp(cor_mat,ave_logrpkm)
}
}
