\name{get_IQR_condition_exp}
\alias{get_IQR_condition_exp}
\title{Calculate the IQR for each bin of the correlation matrix}
\usage{get_IQR_condition_exp(cor_mat, ave_exp)}
\arguments{
\item{cor_mat}{correlation matrix, generated by gene expression matrix, with genes sorted by average expression levels}
\item{ave_exp}{vector, average expression level of each gene for the normalized gene expression matrix}
}
\value{A matrix containing the IQR of each bin}
\description{
  This function allows users to compare the IQRs of a set of 10 by 10 same-size bins that partition the correlation matrix.
}
\examples{
  if(require(spqnData)) {
    data(gtex.4k)
    cor_mat <- cor(t(assay(gtex.4k)))
    ave_logrpkm <- rowData(gtex.4k)$ave_logrpkm
    get_IQR_condition_exp(cor_mat, ave_exp = ave_logrpkm)
}
}
