% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{normalize_correlation}
\alias{normalize_correlation}
\title{Spatial Quantile Normalization}
\usage{
normalize_correlation(cor_mat, ngrp, size_grp, ref_grp)
}
\arguments{
\item{cor_mat}{Matrix, correlation matrix, generated by gene expression matrix, with genes sorted by average expression levels}

\item{ngrp}{Integer, number of bins to use in the estimation of local correlation distributions. To smooth the normalization, SpQN uses rolling bins to estimate the distribution of correlations for genes in each expression level.}

\item{size_grp}{Integer, size of rolling bin, i.e. the number of genes in each edge of the bin.}

\item{ref_grp}{Integer, the location of reference bin that represents target distribution. SpQN uses the reference bin to map the correlations in each disjoint bin to the corresponding percentile in the target distribution.}
}
\value{
The normalized correlation matrix, where the local distribution of correlations are approximately the same across the entire correlation matrix.
}
\description{
The normalize_correlation function allows users to normalize the correlation matrix, such that the distributions of correlations are not dependent of gene expression levels.
}
\examples{
a=rnorm(10000)
b=a/max(abs(a))
cor_ori=array(b,dim=c(100,100))
normalize_correlation(cor_ori,ngrp=10,size_grp=15,ref_grp=9)

}
\name{plot_signal_condition_exp}
\alias{plot_signal_condition_exp}
\title{Plot the signal and background distribution}
\usage{plot_signal_condition_exp(cor_mat, percent_sig) }
\arguments{
\item{cor_mat}{Matrix, correlation matrix, generated by gene expression matrix, with genes sorted by average expression levels}
\item{percent_sig}{Numeric, percentage of signals in each group of the correlation matrix} }
\value{}
\description{This function allows users to visualize the separation between signals and background, conditional on expression levels.
}
\examples{
a=rnorm(10000)
b=a/max(abs(a))
cor_ori=array(b,dim=c(100,100))
plot_signal_condition_exp(cor_ori,0.1)
}
\name{qqplot_condition_exp}
\alias{qqplot_condition_exp}
\title{Qqplot for two bins in the correlation matrix}
\usage{qqplot_condition_exp(cor_mat, i,j)}
\arguments{
\item{cor_mat}{Matrix, correlation matrix, generated by gene expression matrix, with genes sorted by average expression levels}
\item{i}{Integer, bin number of the first bin} }
\item{i}{Integer, bin number of the second bin} }
\value{}
\description{This function allows users to compare the distribution between each bin and the reference bin by qqplot.
}
\examples{
a=rnorm(10000)
b=a/max(abs(a))
cor_ori=array(b,dim=c(100,100))
qqplot_condition_exp(cor_ori,1,9)
}

