% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{normalize_correlation}
\alias{normalize_correlation}
\title{Spatial Quantile Normalization}
\usage{
normalize_correlation(cor_mat, ngrp, size_grp, ref_grp)
}
\arguments{
\item{cor_mat}{Matrix, correlation matrix, generated by gene expression matrix, with genes sorted by average expression levels}

\item{ngrp}{Integer, number of bins to use in the estimation of local correlation distributions. To smooth the normalization, SpQN uses rolling bins to estimate the distribution of correlations for genes in each expression level.}

\item{size_grp}{Integer, size of rolling bin, i.e. the number of genes in each edge of the bin.}

\item{ref_grp}{Integer, the location of reference bin that represents target distribution. SpQN uses the reference bin to map the correlations in each disjoint bins to the corresponding percentile in the target distribution.}
}
\value{
The normalized correlation matrix, where the distribution of correlations are approximately independent of gene expression levels.
}
\description{
The normalize_correlation function allows users to normalize the correlation matrix, such that the local distributions of the correlations are approximately the same.
}
\examples{
a=rnorm(10000)
b=a/max(abs(a))
cor_ori=array(b,dim=c(100,100))
normalize_correlation(cor_ori,ngrp=10,size_grp=15,ref_grp=9)

}
