\name{plot_signal_condition_exp}
\alias{plot_signal_condition_exp}
\title{Plot the signal and background distribution}
\usage{
plot_signal_condition_exp(cor_mat, ave_logrpkm, percent_sig) 
}
\arguments{
\item{cor_mat}{Matrix, correlation matrix, generated by gene expression matrix, with genes sorted by average expression levels}
\item{ave_logrpkm}{Vector, average expression level of each gene}
\item{percent_sig}{Numeric, percentage of signals in each group of the correlation matrix} 
}
\value{A ridge plot visualizing the distribution of background and signals in each group of correlation matrix}
\description{
This function allows users to visualize the separation between signals and background, conditional on expression levels.
}
\examples{
if(require(spqnData)) {
  data(gtex.4k)
  cor_mat <- cor(t(assay(gtex.4k)))
  ave_logrpkm <- rowData(gtex.4k)$ave_logrpkm
  plot_signal_condition_exp(cor_mat,ave_logrpkm,5)}
}
