\name{plot_signal_condition_exp}
\alias{plot_signal_condition_exp}
\title{Plot the signal and background distribution}
\usage{
plot_signal_condition_exp(cor_mat, ave_exp, percent_sig) 
}
\arguments{
\item{cor_mat}{Matrix, correlation matrix, generated by gene expression matrix}
\item{ave_exp}{Vector, average expression level of each gene for the normalized expression matrix}
\item{percent_sig}{Numeric, percentage of signals in each group of the correlation matrix} 
}
\value{A ridge plot visualizing the distribution of background and signals in each bin on the diagonal of correlation matrix}
\description{
This function allows users to visualize the distributions of predicted signals and background, conditional on expression levels. 
The predicted signals are defined by the 0.1% highest correlations in each bin.
}
\examples{
if(require(spqnData)) {
  data(gtex.4k)
  cor_mat <- cor(t(assay(gtex.4k)))
  ave_logrpkm <- rowData(gtex.4k)$ave_logrpkm
  plot_signal_condition_exp(cor_mat, ave_exp=ave_logrpkm, percent_sig=5)}
}
