\name{qqplot_condition_exp}
\alias{qqplot_condition_exp}
\title{Qqplot for two bins in the correlation matrix}
\usage{qqplot_condition_exp(cor_mat, i,j)}
\arguments{
\item{cor_mat}{Matrix, correlation matrix, generated by gene expression matrix, with genes sorted by average expression levels}
\item{i}{Integer, bin number of the first bin}
\item{j}{Integer, bin number of the second bin}
}
\value{}
\description{This function allows users to compare the distribution between each bin and the reference bin by qqplot.
}
\examples{
a=rnorm(10000)
b=a/max(abs(a))
cor_ori=array(b,dim=c(100,100))
qqplot_condition_exp(cor_ori,1,9)
}
