\name{qqplot_condition_exp}
\alias{qqplot_condition_exp}
\title{Q-Q plot for comparing the distribution of two bins in the correlation matrix}
\usage{qqplot_condition_exp(cor_mat,ave_logrpkm, i,j)}
\arguments{
\item{cor_mat}{Matrix, correlation matrix, generated by gene expression matrix, with genes sorted by average expression levels}
\item{ave_logrpkm}{Vector, average expression level of each gene}
\item{i}{Integer, number of the tested bin}
\item{j}{Integer, number of the reference bin}
}
\description{This function allows users to compare the distribution of two bins using qqplot.}
\value{A Q-Q plot that compares the distribution of two bins}
\examples{
a=rnorm(10000)
b=a/max(abs(a))
cor_mat=array(b,dim=c(100,100))
ave_logrpkm=c(1:100)
qqplot_condition_exp(cor_mat,ave_logrpkm,1,9)
}
