\name{qqplot_condition_exp}
\alias{qqplot_condition_exp}
\title{Q-Q plot for comparing the distribution of a bin with the reference bin in the correlation matrix}
\usage{qqplot_condition_exp(cor_mat,ave_logrpkm, i,j)}
\arguments{
\item{cor_mat}{Matrix, correlation matrix, generated by gene expression matrix}
\item{ave_exp}{Vector, average expression level of each gene for the normalized expression matrix}
\item{i}{Integer, row number of the tested bin}
\item{j}{Integer, column number of the tested bin}
}
\description{This function allows users to compare the distribution of a bin with the reference bin using qqplot. There are 
             10 by 10 bins of the same size that partition the correlation matrix. }
\value{A Q-Q plot that compares the distribution of the tested bin with the reference bin, here is the bin in the 9th row 9th 
       column, in the correlation matrix.}
\examples{
if(require(spqnData)) {
  data(gtex.4k)
  cor_mat <- cor(t(assay(gtex.4k)))
  ave_logrpkm <- rowData(gtex.4k)$ave_logrpkm
  qqplot_condition_exp(cor_mat,ave_logrpkm,1,1)}
}
