\name{spqn-package}
\alias{spqn-package}
\alias{spqn}
\docType{package}
\title{
\packageTitle{spqn}
}
\description{
\packageDescription{spqn}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{spqn}
\packageIndices{spqn}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{spqn}

Maintainer: \packageMaintainer{spqn}
}
\references{

}
\keyword{ package }
